consolidate iterator types
